requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.12.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-barcode==0.15.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.7.1
yarl==1.9.2
